# Speed up QM.add_variables_from() and QM.add_linear_from()