Some minor speedups for `QM.add_variables_from()` and `QM.add_linear_from()`

```bash
Commit: c9fb8c3f <main>

qm.TimeAddLinearFrom.time_add_linear_from [desktop/virtualenv-py3.8]
  ok
  ======== ============= =============
  --                  param2          
  -------- ---------------------------
   param1       dict          list    
  ======== ============= =============
    str     1.07±0.05ms   1.10±0.05ms 
    int       332±80μs      295±2μs   
  ======== ============= =============
  started: 2022-04-05 11:00:31, duration: 1.82s

qm.TimeAddVariablesFrom.time_add_variables_from_str [desktop/virtualenv-py3.8]
  ok
  ======== =============
   param1               
  -------- -------------
    str     2.79±0.04ms 
    int     1.36±0.01ms 
  ======== =============
  started: 2022-04-05 11:00:32, duration: 808ms

```

```bash
Commit: d77a246a <feature/speedup-qm-linear-methods>

qm.TimeAddLinearFrom.time_add_linear_from [desktop/virtualenv-py3.8]
  ok
  ======== ============= ===========
  --                 param2         
  -------- -------------------------
   param1       dict         list   
  ======== ============= ===========
    str     1.02±0.01ms   1000±30μs 
    int       305±20μs     271±5μs  
  ======== ============= ===========
  started: 2022-04-05 11:00:03, duration: 1.67s

qm.TimeAddVariablesFrom.time_add_variables_from_str [desktop/virtualenv-py3.8]
  ok
  ======== =============
   param1               
  -------- -------------
    str     1.95±0.01ms 
    int       606±6μs   
  ======== =============
  started: 2022-04-05 11:00:04, duration: 809ms

```